Weatherbot fix + Embed styling
Refactoring

Weatherbot fix

Embeds used everywhere
